break out layer list, re-org files into subfolders
