+++ web/js/components/layer/product-picker/browse/category-cell.js
